chore: reorganise project structure
- source files go into the src directory
- test files (and assets) go into the test directory
- development and documentation assets go into the resources directory
