+++ resources/full.jq
