+++ resources/types.jq
@@ -1,0 +1,1 @@
+to_entries | "TY,Description\n" + (map("\(.key),\(.value)") | join("\n"))
