+++ src/lexer.js
